[MP Rewrite Part 6] Cleanup and fixes (#1512)
* Cleanup and fixes to make code run properly

* Fix test compilation and disable a trapezoidal motion profile edge case.

Some planner tests are still failing. This will be addressed in a later
commit.

* Prettify

* Make e2e planner tests pass.

Because they generate random scenarios, and the RRT planner is not
guaranteed to converge in a finite number of iterations (and the Bezier
approximation may be incorrect), these tests are currently flaky. That
should be fixed before this is merged to staging.

* Tidy fixes

* Linting

* Formatting

* Remove deleted file from CMakeLists, remove reuse condition from settle

* Don't replan in Collect::control

* Optimizations to geometry methods

* Reset and regenerate sim config

This deletes unnecessary variables and moves planning-related variables
into PathPlanner

* Formatting

* Add TODO in EscapeObstacles

* Tidy fixes

* Revert inconsistent clang-tidy recommended fix

* Temporarily disable flaky planner test

* Add NOLINTNEXTLINE

* Remove old sim config and add TODOs for trajectory utils

+++ soccer/planning/DynamicObstacle.hpp
@@ -16,7 +16,7 @@ struct DynamicObstacle {
      * outlive the usage of this object.
      */
     DynamicObstacle(double radius, const Trajectory* path)
-        : circle(Geometry2d::Point(), radius), path(path) {}
+        : circle(Geometry2d::Point(), static_cast<float>(radius)), path(path) {}
     Geometry2d::Circle circle;
     const Trajectory* path;
 };
